repartition of headers
